refactor: improve clarity and correctness in CardTerminal and CardTerminals error handling
- Update type casting for SCARD_E_TIMEOUT and SCARD_W_REMOVED_CARD.
- Simplify logic by replacing switch with if-else structure in CardTerminal.
- Fix initialization order in PcscReaderAdapter constructor.

+++ src/main/cpp/CardTerminals.cpp
@@ -71,7 +71,7 @@ CardTerminals::waitForChange(long timeout)
 
     LONG rv = SCardGetStatusChange(
         mContext, timeout, mKnownReaders.data(), static_cast<DWORD>(mKnownReaders.size()));
-    if (rv == SCARD_E_TIMEOUT) {
+    if (rv == static_cast<LONG>(SCARD_E_TIMEOUT)) {
         return false;
     }
